Create package.json
File states that its homepage is "github.com/oscillate123", this is because I had a friend help with an issue in the code where it was uploaded to his GitHub (via an external program called SourceTree) so he could examine it. 
Because of this the package.json file states that it's homepage is his GitHub account for some reason.